Use sauce namespace for commands
