+++ lib/commands/index.js
@@ -1,6 +1,6 @@
 /* jshint node: true */
 module.exports = {
   'sauce': require('./sauce'),
-  'start-sauce-connect': require('./start-sauce-connect'),
-  'stop-sauce-connect': require('./stop-sauce-connect')
+  'sauce:connect': require('./sauce-connect'),
+  'sauce:disconnect': require('./sauce-disconnect')
 };
