+++ lib/commands/sauce-disconnect.js
@@ -4,13 +4,14 @@
 var runScript = require('../utils/run-script');
 
 module.exports = {
-  name: 'start-sauce-connect',
+  name: 'sauce:disconnect',
+  aliases: ['stop-sauce-connect'],
   works: 'insideProject',
-  description: 'Start a sauce connect tunnel.',
+  description: 'Disconnect from sauce.',
 
   availableOptions: [],
 
   run: function() {
-    return runScript(this.ui, __dirname + '/../../vendor/start-sauce-connect.sh');
+    return runScript(this.ui, __dirname + '/../../vendor/stop-sauce-connect.sh');
   }
 };
